chore: move `operators` to the file `operators.ts` and rename `constants.ts` to `regexps.ts`
